+++ examples/example3.hs
@@ -1,13 +1,16 @@
+{-# LANGUAGE FlexibleContexts #-}
 module Main where
 
 import Prelude hiding (length, sum)
 import Control.Applicative ((<$>))
 import Data.NetCDF
 import Data.NetCDF.HMatrix
+import Data.NetCDF.Store (smap)
 import qualified Data.Map as M
 import Foreign.C
 import Foreign.Storable
-import Numeric.Container
+import Numeric.LinearAlgebra
+import Numeric.LinearAlgebra.Devel (foldVector)
 
 type VRet a = IO (Either NcError (HVector a))
 type MRet a = IO (Either NcError (HMatrix a))
@@ -32,21 +35,25 @@ main = do
   -- Get longitude variable: metadata, then all values, then a slice
   -- of values.
   let (Just lonvar) = ncVar nc "longitude"
-  Right (HVector lon) <- get nc lonvar :: VRet CFloat
-  let mlon = mean lon
+  Right lon' <- get nc lonvar :: VRet CFloat
+  let HVector lon = smap realToFrac lon' :: HVector Float
+      mlon = mean lon
   putStrLn $ "longitude: " ++ show lon ++ " -> " ++ show mlon
-  Right (HVector lon2) <- getS nc lonvar [0] [72] [2] :: VRet CFloat
-  let mlon2 = mean lon2
+  Right lon2' <- getS nc lonvar [0] [72] [2] :: VRet CFloat
+  let HVector lon2  = smap realToFrac lon2' :: HVector Float
+      mlon2 = mean lon2
   putStrLn $ "longitude (every 2): " ++ show lon2 ++ " -> " ++ show mlon2
 
   -- Get latitude variable: metadata, then all values, then a slice of
   -- values.
   let (Just latvar) = ncVar nc "latitude"
-  Right (HVector lat) <- get nc latvar :: VRet CFloat
-  let mlat = mean lat
+  Right lat' <- get nc latvar :: VRet CFloat
+  let HVector lat  = smap realToFrac lat' :: HVector Float
+      mlat = mean lat
   putStrLn $ "latitude: " ++ show lat ++ " -> " ++ show mlat
-  Right (HVector lat2) <- getS nc latvar [0] [37] [2] :: VRet CFloat
-  let mlat2 = mean lat2
+  Right lat2' <- getS nc latvar [0] [37] [2] :: VRet CFloat
+  let (HVector lat2)  = smap realToFrac lat2' :: HVector Float
+      mlat2 = mean lat2
   putStrLn $ "latitude (every 2): " ++ show lat2 ++ " -> " ++ show mlat2
 
   -- Get Z500 variable: get a slice, do COARDS short -> double
@@ -59,10 +66,11 @@ main = do
       slice1 = cmap ((/ 9.8) . realToFrac) slice1tmp2 :: Matrix Double
   putStrLn $ "size slice1 = " ++
     show (rows slice1) ++ " x " ++ show (cols slice1)
-  putStrLn $ "lon(i=25) = " ++ show (lon @> (25 - 1))
-  putStrLn $ "lat(j=40) = " ++ show (lat @> (nlat - 40))
-  let v @!! (i, j) = v @@> (nlat - i, j - 1)
+  putStrLn $ "lon(i=25) = " ++ show (lon `atIndex` (25 - 1))
+  putStrLn $ "lat(j=40) = " ++ show (lat `atIndex` (nlat - 40))
+  let (@!!) :: Matrix Double -> (Int, Int) -> Double
+      v @!! (i, j) = v `atIndex` (nlat - i, j - 1)
   putStrLn $ "slice1(i=25,j=40) = " ++ show (slice1 @!! (25, 40))
 
-mean :: (Storable a, Fractional a) => Vector a -> a
-mean xs = (foldVector (+) 0 xs) / fromIntegral (dim xs)
+mean :: (Storable a, Fractional a, Container Vector a) => Vector a -> a -- FlexibleContexts
+mean xs = (foldVector (+) 0 xs) / fromIntegral (size xs)